requirements/base.txt
# Base
pytz==2021.3
python-slugify==5.0.2
pillow==9.0.0
psycopg2-binary==2.9.3

# Django
django>=4.0.1

# Django REST Framework
djangorestframework==3.13.1
django-filter==21.1

# JWT
pyjwt==2.3.0

# Cors
django-cors-headers==3.11.0

# Passwords security
argon2-cffi==21.3.0

# Static files
whitenoise==5.3.0

# NYSE Symbols
finnhub-python==2.4.8
